fix wrong function type

+++ rvp64_scan/builtin-rvp64-insb.c
@@ -9,7 +9,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-uint32_t bsni (uint32_t ra, uint32_t rb)
+uint64_t bsni (uint64_t ra, uint32_t rb)
 {
   return __rv__insb (ra, rb, 1);
 }
