+++ rvp64_scan/builtin-rvp64-kslra32.c
@@ -9,7 +9,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-uint32_t arlsk (uint32_t ra, uint32_t rb)
+uint64_t arlsk (uint64_t ra, uint32_t rb)
 {
   return __rv__kslra32 (ra, rb);
 }
