+++ rvp64_scan/builtin-rvp64-srai32.c
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-uint32_t iars (uint32_t ra)
+uint64_t iars (uint64_t ra)
 {
   return __rv__sra32 (ra, 4);
 }
