+++ rvp64_scan/builtin-rvp64-srli16.c
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-uint32_t ilrs (uint32_t ra)
+uint64_t ilrs (uint64_t ra)
 {
   return __rv__srl16 (ra, 4);
 }
